Support Symfony 6
