+++ composer.json
@@ -23,9 +23,9 @@
         "doctrine/doctrine-bundle": "^2.4",
         "illuminate/collections": "^8.47",
         "meilisearch/meilisearch-php": "^0.20",
-        "symfony/filesystem": "^4.0 || ^5.0",
-        "symfony/property-access": "^4.0 || ^5.0",
-        "symfony/serializer": "^4.0 || ^5.0"
+        "symfony/filesystem": "^4.0 || ^5.0 || ^6.0",
+        "symfony/property-access": "^4.0 || ^5.0 || ^6.0",
+        "symfony/serializer": "^4.0 || ^5.0 || ^6.0"
     },
     "require-dev": {
         "doctrine/orm": "^2.9",
@@ -38,10 +38,10 @@
         "phpstan/phpstan-symfony": "^0.12.37",
         "phpunit/php-code-coverage": "^9.2",
         "phpunit/phpunit": "^9.5",
-        "symfony/doctrine-bridge": "^4.0 || ^5.0",
-        "symfony/http-client": "^4.0 || ^5.0",
-        "symfony/phpunit-bridge": "^4.0 || ^5.0",
-        "symfony/yaml": "^4.0 || ^5.0"
+        "symfony/doctrine-bridge": "^4.0 || ^5.0 || ^6.0",
+        "symfony/http-client": "^4.0 || ^5.0 || ^6.0",
+        "symfony/phpunit-bridge": "^4.0 || ^5.0 || ^6.0",
+        "symfony/yaml": "^4.0 || ^5.0 || ^6.0"
     },
     "autoload": {
         "psr-4": {
